TERMINATING A SPECIFIC PROCESS
From Windows PowerShell, you can use Get-Process to search for a process by name. The "totally_not_malicious" process is running on this machine, too. Search for it, using this command:Get-Process -Name "totally_not_malicious"
Each row represents a process, and one of the columns shows the process ID,in the picture